requirements.txt
# sudo apt-get install libmysqlclient-dev 先装mysql 才能成功
# sudo apt-get install libsqlite3-dev  先装再编译python不然ipython 无效
Django==3.0
PyYAML==5.4.1
mysqlclient==2.0.3
djangorestframework==3.12.4
Pillow==8.1.2
pypinyin==0.41.0
requests==2.25.1
django-import-export==2.5.0
wechatpy==1.8.15
pysnooper==0.4.3
redis==3.5.3
pycryptodome==3.10.1
python-decouple==3.1
django-admin-rangefilter==0.8.1
django-filter==2.4.0
django-nested-admin==3.3.3
qrcode==5.3
pydantic==1.8.1
schedule==0.5.0
zerorpc==0.6.3
# djantic==0.2.0
pyzmq==22.0.3
hashids==1.3.1
psycopg2-binary==2.9.1
# tencentcloud-sdk-python==3.0.240
requests-toolbelt==0.9.1
simplejson==3.17.5
django_jsonfield_backport==1.0.4
django-jet==1.0.8
user-agents==2.2.0
gunicorn==20.1.0
django-cors-headers==2.1.0
xlsxwriter==0.9.6
sqlparse==0.2.2
unicodecsv==0.14.1
xlwt==1.2.0
captcha==0.3
django-q
openpyxl==3.0.10
alibabacloud_dysmsapi20170525==2.0.24
pycryptodomex
celery==4.4.0
importlib-metadata==4.8.3
#django-celery-results==1.2.1
#flower
ipython
cn2an
# pip install appium-python-client==1.3.0 -i https://pypi.tuna.tsinghua.edu.cn/simple
appium-python-client==1.3.0
pyautogui==0.9.53
django-celery-beat==2.1.0
django-redis==5.0.0
pyjwt==1.7.1
selenium==3.141.0
alibabacloud_dytnsapi20200217==2.8.1
tencentcloud-sdk-python==3.0.1439
aioredis==2.0.1
orjson
coreapi